+++ apps/indexer/test/indexer/temporary/uncataloged_token_transfers_test.exs
@@ -34,20 +34,23 @@ defmodule Indexer.Temporary.UncatalogedTokenTransfersTest do
       block = insert(:block)
       address = insert(:address)
 
-      log =
-        insert(:token_transfer_log,
-          transaction:
-            insert(:transaction,
-              block_number: block.number,
-              block_hash: block.hash,
-              cumulative_gas_used: 0,
-              gas_used: 0,
-              index: 0
-            ),
-          address_hash: address.hash
+      transaction =
+        insert(:transaction,
+          block_number: block.number,
+          block_hash: block.hash,
+          cumulative_gas_used: 0,
+          gas_used: 0,
+          index: 0
         )
 
-      block_number = log.transaction.block_number
+      log =
+        insert(:token_transfer_log,
+          transaction: transaction,
+          address_hash: address.hash,
+          block: block
+        )
+
+      block_number = log.block_number
 
       expected_state = %{task_ref: nil, block_numbers: [block_number], retry_interval: 1}
       state = %{task_ref: nil, block_numbers: [], retry_interval: 1}